Background sending to a Kubernetes-based RabbitMQ
The normal oslo.messaging retry logic does not work when RabbitMQ is
run as a Kubernetes service.

This is because the K8S service proxy does not always close the socket
when the RabbitMQ container dies, and the oslo.messaging logic expects
a closed socket before retrying.

Therefore we introduce here another thread to handle the actual event
sending, separate to the queue processing thread. The queue processing
thread waits for the event sending thread to indicate success, but in
the event of a timeout it will spawn a new event-sending thread which
will establish a new socket to the K8S service proxy and its new
RabbitMQ container.

Change-Id: I223eaad08bd22744dbd304e3fd3ecceaba347b6e